Remove `/Res` in .gitignore
I know that some of you are strongly opposed to putting binary files in
the github repository, but NOW I PUT TEXTURES IN IT. This is better than
putting them in BaiduYun disk for it cannot do version control (and
updating will be troublesome). Also, the free LFS service is even worse
beacuse the quota is not enough for use.

I knew there is another Minecraft-like game on GitHub called Minetest
(https://www.github.com/minetest/minetest). You can see there are PNG
textures in their repo. You can try to clone that. It is not very slow. I
think we can also put compressed textures in our repo.

According to @Miigon 's idea, put textures outside the repo may cause
some troubles, for example, you downloaded the textures but they are not
in the correct version. In a word, PLEASE PUT TEXTURES AND OTHER RESOURCES
IN THE REPO... ALSO DON'T CRITICIZE PLEASE...